requirements.txt
boto3==1.38.13
botocore==1.34.69
Flask==3.1.0
flask_cors==5.0.1
langchain==0.3.25
langchain_core==0.3.59
langchain_google_genai==2.1.4
nbformat==5.10.4
openai==1.78.0
protobuf==3.20.3
python-dotenv==1.1.0
Requests==2.32.3
streamlit==1.45.0
